create visualize/ package; convert util/visualize.py and lit/plot.py to deprecation shims
