+++ src/rfdetr/visualize/__init__.py
@@ -1,0 +1,15 @@
+# ------------------------------------------------------------------------
+# RF-DETR
+# Copyright (c) 2025 Roboflow. All Rights Reserved.
+# Licensed under the Apache License, Version 2.0 [see LICENSE for details]
+# ------------------------------------------------------------------------
+
+"""Visualization utilities for RF-DETR."""
+
+from rfdetr.visualize.data import save_gt_predictions_visualization
+from rfdetr.visualize.training import plot_metrics
+
+__all__ = [
+    "plot_metrics",
+    "save_gt_predictions_visualization",
+]
